Fuzzy controller chrasher ikke længer
Fuck floats... Mangler noget tuning af controlleren, muligvis nogle regler, måske det løses andet sted i koden, det skal besluttes

+++ marbel_controller.h
@@ -16,8 +16,8 @@ class marbel_Controller
 public:
     marbel_Controller();
     virtual ~marbel_Controller() = default;
-    virtual float buildController(float cent);
-    virtual ControlOutput getControlOutput(float cent);
+    virtual float buildController(int cent);
+    virtual ControlOutput getControlOutput(int cent);
 protected:
     //LaserScanner*        m_pcLaserScanner;
